geom_tufteboxplot {ggthemes}	R Documentation

Tufte's Box Plot

Description

Edward Tufte's revisions of the box plot as described in The Visual Display of Quantitative Information.

Usage

geom_tufteboxplot(mapping = NULL, data = NULL, stat = "fivenumber",
  position = "dodge", outlier.colour = "black", outlier.shape = 19,
  outlier.size = 1.5, outlier.stroke = 0.5, voffset = 0.01,
  hoffset = 0.005, na.rm = FALSE, show.legend = NA, inherit.aes = TRUE,
  median.type = "point", whisker.type = "line", ...)

Arguments

mapping
Set of aesthetic mappings created by aes or aes_. If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot.

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame., and will be used as the layer data.

stat
The statistical transformation to use on the data for this layer, as a string.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

outlier.colour
colour for outlying points

outlier.shape
shape of outlying points

outlier.size
size of outlying points

outlier.stroke
stroke for outlying points

voffset
controls the size of the gap in the line representing the median when median.type = 'line'. This is a fraction of the range of y.

hoffset
controls how much the interquartile line is offset from the whiskers when median.type = 'line'. This is a fraction of the range of x.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders.

median.type
If 'point', then the median is represented by a point, and the interquartile range by a gap in the line. If median.type='line', then the interquartile range is represented by a line, possibly offset, and the median by a gap in the line.

whisker.type
If 'line', then whiskers are represented by lines. If 'point', then whiskers are represented by points at ymin and ymax.

...
other arguments passed on to layer. These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3. They may also be parameters to the paired geom/stat.

Aesthetics

x [required]

y [required]

colour

size

linetype

shape

fill

alpha

References

Tufte, Edward R. (2001) The Visual Display of Quantitative Information, Chapter 6.

McGill, R., Tukey, J. W. and Larsen, W. A. (1978) Variations of box plots. The American Statistician 32, 12-16.

See Also

geom_boxplot

Other geom tufte: geom_rangeframe

Examples

### Keywords: datasets

library("ggplot2")
p <- ggplot(mtcars, aes(factor(cyl), mpg))
## with a point for the median and lines for whiskers
p + geom_tufteboxplot()
## with a line for the interquartile range and points for whiskers
p + geom_tufteboxplot(median.type = "line", whisker.type = "point", hoffset = 0)
## with a wide line for the interquartile range and lines for whiskers
p + geom_tufteboxplot(median.type = "line", hoffset = 0, width = 3)
## with an offset line for the interquartile range and lines for whiskers
p + geom_tufteboxplot(median.type='line')

[Package ggthemes version 3.4.2 Index]