ggthemes {ggthemes}	R Documentation

ggthemes

Description

This package contains extra themes, scales, and geoms, and functions for and related to ggplot2.

Details

In addition to the help pages, see the README page on github for examples.

[Package ggthemes version 3.4.2 Index]