scale_colour_economist {ggthemes}	R Documentation

Economist color scales

Description

Color scales using the colors in the Economist graphics.

Usage

scale_colour_economist(stata = FALSE, ...)

scale_color_economist(stata = FALSE, ...)

scale_fill_economist(stata = FALSE, ...)

Arguments

stata
Use the palette in the Stata economist scheme.

...
Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.

See Also

theme_economist for examples.

Other colour economist: economist_pal

[Package ggthemes version 3.4.2 Index]