scale_colour_fivethirtyeight {ggthemes}	R Documentation

fivethirtyeight.com color scales

Description

Color scales using the colors in the fivethirtyeight graphics.

Usage

scale_colour_fivethirtyeight(...)

scale_color_fivethirtyeight(...)

scale_fill_fivethirtyeight(...)

Arguments

...
Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.

See Also

theme_fivethirtyeight for examples.

Other colour fivethirtyeight: fivethirtyeight_pal

[Package ggthemes version 3.4.2 Index]