scale_color_pander {ggthemes}	R Documentation

Color scale from the pander package

Description

The pander ships with a default colorblind and printer-friendly color palette borrowed from http://jfly.iam.u-tokyo.ac.jp/color/.

Usage

scale_color_pander(...)

scale_colour_pander(...)

scale_fill_pander(...)

Arguments

...
Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.

See Also

theme_pander

Other colour pander: palette_pander

[Package ggthemes version 3.4.2 Index]