scale_shape_tremmel {ggthemes}	R Documentation

Shape scales from Tremmel (1995)

Description

Shape scales from Tremmel (1995)

Usage

scale_shape_tremmel(overlap = FALSE, n3alt = TRUE, ...)

Arguments

overlap
use an empty circle instead of a solid circle when n == 2.

n3alt
If TRUE then use a solid circle, plus sign and empty triangle, else use a solid circle, empty circle, and empty triangle.

...
common discrete scale parameters: name, breaks, labels, na.value, limits and guide. See discrete_scale for more details

See Also

tremmel_shape_pal for a description of the palette.

Other shapes: circlefill_shape_pal, cleveland_shape_pal, scale_shape_circlefill, scale_shape_cleveland, tremmel_shape_pal

Examples

library("ggplot2")
(ggplot(mtcars, aes(x=mpg, y=hp, shape=factor(cyl)))
 + geom_point() + scale_shape_tremmel())
(ggplot(mtcars, aes(x=mpg, y=hp, shape=factor(cyl)))
 + geom_point() + scale_shape_tremmel(n3alt=FALSE))
(ggplot(mtcars, aes(x=mpg, y=hp, shape=factor(am)))
 + geom_point() + scale_shape_tremmel())
(ggplot(mtcars, aes(x=mpg, y=hp, shape=factor(am)))
 + geom_point() + scale_shape_tremmel(overlap=TRUE))

[Package ggthemes version 3.4.2 Index]