scale_colour_wsj {ggthemes}	R Documentation

Wall Street Journal color and fill scales

Description

Colour and fill scales which use the palettes in wsj_pal and are meant for use with theme_wsj.

Usage

scale_colour_wsj(palette = "colors6", ...)

scale_color_wsj(palette = "colors6", ...)

scale_fill_wsj(palette = "colors6", ...)

Arguments

palette
character The color palette to use. This must be a name in ggthemes_data$wsj$palettes.

...
Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.

See Also

Other colour wsj: wsj_pal

[Package ggthemes version 3.4.2 Index]