smart_digits {ggthemes}	R Documentation

Format numbers with automatic number of digits

Description

Format numbers with automatic number of digits

Usage

smart_digits(x, ...)

smart_digits_format(x, ...)

Arguments

x
A numeric vector to format

...
Parameters passed to format

Value

smart_digits returns a character vector. smart_digits_format returns a function with a single argument x, a numeric vector, that returns a charactger vector.

Author(s)

Josh O'Brien, Baptise Auguie, Jeffrey B. Arnold

References

Josh O'Brien, http://stackoverflow.com/questions/23169938/select-accuracy-to-display-additional-axis-breaks/23171858#23171858.

[Package ggthemes version 4.0.0 Index]