stat_fivenumber {ggthemes}	R Documentation

Calculate components of a five-number summary

Description

The five number summary of a sample is the minimum, first quartile, median, third quartile, and maximum.

Usage

stat_fivenumber(mapping = NULL, data = NULL, geom = "boxplot", qs = c(0,
  0.25, 0.5, 0.75, 1), na.rm = FALSE, position = "identity",
  show.legend = NA, inherit.aes = TRUE, ...)

Arguments

mapping
Set of aesthetic mappings created by aes or aes_. If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot.

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame., and will be used as the layer data.

geom
The geometric object to use display the data

qs
Quantiles to use for the five number summary.

na.rm
If FALSE (the default), removes missing values with a warning. If TRUE silently removes missing values.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders.

...
other arguments passed on to layer. These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3. They may also be parameters to the paired geom/stat.

Value

A data frame with additional columns:

width
width of boxplot

min
minimum

lower
lower hinge, 25% quantile

middle
median, 50% quantile

upper
upper hinge, 75% quantile

max
maximum

See Also

stat_boxplot

[Package ggthemes version 3.4.2 Index]