tableau_color_pal {ggthemes}	R Documentation

Color Palettes based on Tableau (discrete)

Description

Color palettes used in Tableau.

Usage

tableau_color_pal(palette = "tableau10")

Arguments

palette
Palette name.

References

http://vis.stanford.edu/color-names/analyzer/

Maureen Stone, 'Designing Colors for Data' (slides), at the International Symposium on Computational Aesthetics in Graphics, Visualization, and Imaging, Banff, AB, Canada, June 22, 2007 http://www.stonesc.com/slides/CompAe%202007.pdf.

Heer, Jeffrey and Maureen Stone, 2012 'Color Naming Models for Color Selection, Image Editing and Palette Design', ACM Human Factors in Computing Systems (CHI) http://vis.stanford.edu/files/2012-ColorNameModels-CHI.pdf.

See Also

Other colour tableau: scale_colour_gradient2_tableau, scale_colour_gradient_tableau, scale_colour_tableau, tableau_div_gradient_pal, tableau_seq_gradient_pal

Examples

library("scales")
show_col(tableau_color_pal('tableau20')(20))
show_col(tableau_color_pal('tableau10')(10))
show_col(tableau_color_pal('tableau10medium')(10))
show_col(tableau_color_pal('tableau10light')(10))
show_col(tableau_color_pal('colorblind10')(10))
show_col(tableau_color_pal('trafficlight')(10))
show_col(tableau_color_pal('purplegray12')(12))
show_col(tableau_color_pal('bluered12')(12))
show_col(tableau_color_pal('greenorange12')(12))
show_col(tableau_color_pal('cyclic')(20))

[Package ggthemes version 3.4.2 Index]