theme_clean {ggthemes}	R Documentation

Clean ggplot theme

Description

Clean ggplot theme with no panel background, black axis lines and grey fill colour for chart elements.

Usage

theme_clean(base_size = 12, base_family = "sans")

Arguments

base_size
Base font size.

base_family
Base font family.

Author(s)

Konrad Zdeb name.surname@me.com

See Also

Other themes: theme_base, theme_foundation, theme_igray, theme_par, theme_solid

Examples

library("ggplot2")
p <- ggplot(mtcars, aes(x = wt, y = mpg, colour = factor(gear))) +
  geom_point() +
  facet_wrap(~am) +
  geom_smooth(method = "lm", se = FALSE) +
  theme_clean()
p

[Package ggthemes version 4.2.0 Index]