theme_par {ggthemes}	R Documentation

Theme which takes its values from the current ‘base’ graphics parameter values in par.

Description

Currently this theme uses the values of the parameters: "code", ""ps"", "code" "family", "fg", "bg", "adj", "font", "cex.axis", "cex.lab", "cex.main", "cex.sub", "col.axis", "col.lab", "col.main", "col.sub", "font", "font.axis", "font.lab", "font.main", "font.sub", "las", "lend", "lheight", "lty", "mar", "ps", "tcl", "tck", "xaxt", "yaxt".

Usage

theme_par(base_size = par()$ps, base_family = par()$family)

Arguments

base_size
base font size

base_family
base font family

Details

This theme does not translate the base graphics perfectly, so the graphs produced by it will not be identical to those produced by base graphics, most notably in the spacing of the margins.

See Also

Other themes: theme_base, theme_foundation, theme_igray, theme_solid

Examples

###   parameter values in 'par'.

library("ggplot2")
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
     colour=factor(gear))) + facet_wrap(~am)
par(font = 2, col.lab = "red", fg = "blue")
p + theme_par()

[Package ggthemes version 3.4.2 Index]