economist_pal {ggthemes}	R Documentation

Economist color palette (discrete)

Description

The hues in the palette are blues, grays, and greens. Red is not included in these palettes and should be used to indicate important data.

Usage

economist_pal(stata = FALSE, fill = TRUE)

Arguments

stata
Use the palette in the Stata economist scheme.

fill
Use the fill palette.

See Also

Other colour economist: scale_colour_economist

Examples

library(scales)
show_col(economist_pal()(6))
## fill palette
show_col(economist_pal(fill=TRUE)(6))
## RGB values from Stata's economist scheme
show_col(economist_pal(stata=TRUE)(16))

[Package ggthemes version 3.4.2 Index]