excel_new_pal {ggthemes}	R Documentation

Excel (current versions) color palettes (discrete)

Description

Color palettes used by current versions of Microsoft Office and Excel.

Usage

excel_new_pal(theme = "Office Theme")

Arguments

theme
The name of the Office theme or color theme (not to be confused with ggplot2 themes) from which to derive the color palette. Available themes include: "Atlas", "Badge", "Berlin", "Celestial", "Crop", "Depth", "Droplet", "Facet", "Feathered", "Gallery", "Headlines", "Integral", "Ion Boardroom", "Ion", "Madison", "Main Event", "Mesh", "Office Theme", "Organic", "Parallax", "Parcel", "Retrospect", "Savon", "Slice", "Vapor Trail", "View", "Wisp", "Wood Type", "Aspect", "Blue Green", "Blue II", "Blue Warm", "Blue", "Grayscale", "Green Yellow", "Green", "Marquee", "Median", "Office 2007-2010", "Orange Red", "Orange", "Paper", "Red Orange", "Red Violet", "Red", "Slipstream", "Violet II", "Violet", "Yellow Orange", "Yellow"

See Also

Other colour excel: excel_pal, scale_colour_excel_new, scale_fill_excel

Examples

library("scales")

for (i in names(ggthemes::ggthemes_data$excel$palettes)) {
  show_col(excel_new_pal(theme = i))(6)
}

[Package ggthemes version 4.2.0 Index]