Highcharts color palette (discrete)

The Highcharts uses many different color palettes in its plots. This collects a few of them.

Usage

hc_pal(palette = "default")

Arguments

palette

character The name of the Highcharts theme to use. One of "default", or "darkunica".

See also