hc_pal {ggthemes}	R Documentation

Highcharts JS color palette (discrete)

Description

The Highcharts JS uses many different color palettes in its plots. This collects a few of them.

Usage

hc_pal(palette = "default")

Arguments

palette
character The name of the Highcharts theme to use.

Palettes

The following palettes are defined,

default

dark-unica

See Also

Other colour hc: scale_colour_hc

[Package ggthemes version 4.0.0 Index]