Color palette from the pander package

The pander ships with a default colorblind and printer-friendly color palette borrowed from https://jfly.iam.u-tokyo.ac.jp/color/.

Usage

palette_pander(n, random_order = FALSE)

Arguments

n

number of colors. This palette supports up to eight colors.

random_order

if the palette should be reordered randomly before rendering each plot to get colorful images

See also

Examples

if (FALSE) { # \dontrun{
  palette_pander(TRUE)
} # }