palette_pander {ggthemes}	R Documentation

Color palette from the pander package

Description

The pander ships with a default colorblind and printer-friendly color palette borrowed from http://jfly.iam.u-tokyo.ac.jp/color/.

Usage

palette_pander(n, random_order = FALSE)

Arguments

n
number of colors

random_order
if the palette should be reordered randomly before rendering each plot to get colorful images

See Also

Other colour pander: scale_color_pander

Examples

## Not run: 
palette_pander(TRUE)

## End(Not run)

[Package ggthemes version 3.4.2 Index]