scale_colour_hc {ggthemes}	R Documentation

Highcharts color and fill scales

Description

Colour and fill scales which use the palettes in hc_pal and are meant for use with theme_hc.

Usage

scale_colour_hc(palette = "default", ...)

scale_color_hc(palette = "default", ...)

scale_fill_hc(palette = "default", ...)

Arguments

palette
character The color palette to use. This must be a name in ggthemes_data$hc$palettes.

...
Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.

[Package ggthemes version 3.4.2 Index]