scale_shape_cleveland {ggthemes}	R Documentation

Shape scales from Cleveland "Elements of Graphing Data"

Description

Shape scales from Cleveland "Elements of Graphing Data"

Usage

scale_shape_cleveland(overlap = TRUE, ...)

Arguments

overlap
logical Use the scale for overlapping points?

...
Arguments passed on to discrete_scale

palette

A palette function that when called with a single integer argument (the number of levels in the scale) returns the values that they should take.

breaks

One of:

NULL for no breaks

waiver() for the default breaks computed by the transformation object

A character vector of breaks

A function that takes the limits as input and returns breaks as output

limits

A character vector that defines possible values of the scale and their order.

drop

Should unused factor levels be omitted from the scale? The default, TRUE, uses the levels that appear in the data; FALSE uses all the levels in the factor.

na.translate

Unlike continuous scales, discrete scales can easily show missing values, and do so by default. If you want to remove missing values from a discrete scale, specify na.translate = FALSE.

na.value

If na.translate = TRUE, what value aesthetic value should missing be displayed as? Does not apply to position scales where NA is always placed at the far right.

aesthetics

The names of the aesthetics that this scale works with

scale_name

The name of the scale

name

The name of the scale. Used as the axis or legend title. If waiver(), the default, the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.

labels

One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

A function that takes the breaks as input and returns labels as output

guide

A function used to create a guide or its name. See guides() for more info.

super

The super class to use for the constructed scale

References

Cleveland WS. The Elements of Graphing Data. Revised Edition. Hobart Press, Summit, NJ, 1994, pp. 154-164, 234-239.

See Also

cleveland_shape_pal() for a description of the palette.

Other shapes: circlefill_shape_pal, cleveland_shape_pal, scale_shape_circlefill, scale_shape_tremmel, tremmel_shape_pal

[Package ggthemes version 4.2.0 Index]