scale_shape_few {ggthemes}	R Documentation

Scales for shapes from "Show Me the Numbers"

Description

scale_shape_few maps discrete variables to five easily discernible shapes. It is based on the shape palette suggested in Few (2012).

Usage

scale_shape_few(...)

Arguments

...
Common discrete_scale parameters. See discrete_scale for more details.

References

Few, S. (2012) Show Me the Numbers: Designing Tables and Graphs to Enlighten, Analytics Press, p. 208.

See Also

scale_shape_few for the shape palette that this scale uses.

[Package ggthemes version 4.0.0 Index]