solarized_pal {ggthemes}	R Documentation

Solarized color palette (discrete)

Description

Qualitative color palate based on the Ethan Schoonover's Solarized palette, http://ethanschoonover.com/solarized. This palette supports up to seven values.

Usage

solarized_pal(accent = "blue")

Arguments

accent
character Starting color.

Note

For a given starting color and number of colors in the palette, the other colors are the combination of colors that maximizes the total Euclidean distance between colors in L*a*b space.

See Also

Other solarized colour: scale_fill_solarized

Examples

library("scales")

show_col(solarized_pal()(2))
show_col(solarized_pal()(3))
show_col(solarized_pal("red")(4))

[Package ggthemes version 4.2.0 Index]