tableau_shape_pal {ggthemes}	R Documentation

Tableau Shape Palettes (discrete)

Description

Shape palettes used by Tableau.

Usage

tableau_shape_pal(palette = c("default", "filled", "proportions"))

Arguments

palette
Palette name.

Details

Not all shape palettes in Tableau are supported. Additionally, these palettes are not exact, and use the best unicode character for the shape palette.

Since these palettes use unicode characters, their look may depend on the font being used, and not all characters may be available.

Shape palettes in Tableau are used to expose images for use a markers in charts, and thus are sometimes groupings of closely related symbols.

See Also

Other shape tableau: scale_shape_tableau

Examples

## Not run: 
  # need to set a font containing these values
  show_shapes(tableau_shape_pal()(5))

## End(Not run)

[Package ggthemes version 4.2.0 Index]