theme_foundation {ggthemes}	R Documentation

Foundation Theme

Description

This theme is designed to be a foundation from which to build new themes, and not meant to be used directly. theme_foundation() is a complete theme with only minimal number of elements defined. It is easier to create new themes by extending this one rather than theme_gray() or theme_bw(), because those themes define elements deep in the hierarchy.

Usage

theme_foundation(base_size = 12, base_family = "")

Arguments

base_size
base font size

base_family
base font family

Details

This theme takes theme_gray() and sets all colour and fill values to NULL, except for the top-level elements (line, rect, and title), which have colour = "black", and fill = "white". This leaves the spacing and-non colour defaults of the default ggplot2 themes in place.

See Also

Other themes: theme_base, theme_clean, theme_igray, theme_par, theme_solid

[Package ggthemes version 4.2.0 Index]