theme_igray {ggthemes}	R Documentation

Inverse gray theme

Description

Theme with white panel and gray background.

Usage

theme_igray(base_size = 12, base_family = "")

Arguments

base_size
base font size

base_family
base font family

Details

This theme inverts the colors in the theme_gray, a white panel and a light gray area around it. This keeps a white background for the color scales like theme_bw. But by using a gray background, the plot is closer to the typographical color of the document, which is the motivation for using a gray panel in theme_gray. This is similar to the style of plots in Stata and Tableau.

See Also

theme_gray, theme_bw

Other themes: theme_base, theme_foundation, theme_par, theme_solid

Examples

library("ggplot2")
p <- ggplot(mtcars) +
    geom_point(aes(x = wt, y = mpg, colour=factor(gear))) +
    facet_wrap(~am)
p + theme_igray()

[Package ggthemes version 3.4.2 Index]