wsj_pal {ggthemes}	R Documentation

Wall Street Journal color palette (discrete)

Description

The Wall Street Journal uses many different color palettes in its plots. This collects a few of them, but is by no means exhaustive. Collections of these plots can be found on the WSJ Graphics Twitter feed and Pinterest.

Usage

wsj_pal(palette = "colors6")

Arguments

palette
character The color palette to use: . "rgby", "red_green", "black_green", "dem_rep", "colors6"

Palettes

The following palettes are defined,

rgby

Red/Green/Blue/Yellow theme. Examples: https://twitpic.com/b2e3v2. Up to four values.

red_green

Green/red two-color scale for good/bad. Examples: https://twitpic.com/b1avj6, http://twitpic.com/a4kxcl.

green_black

Black-green 4-color scale for 'Very negative', 'Somewhat negative', 'somewhat positive', 'very positive'. Examples: https://twitpic.com/awbua0.

dem_rep

Democrat/Republican/Undecided blue/red/gray scale. Examples: https://twitpic.com/awbua0.

colors6

Red, blue, gold, green, orange, and black palette. Examples: https://twitpic.com/9gfg5q.

See Also

Other colour wsj: scale_colour_wsj

[Package ggthemes version 4.0.0 Index]